requirements.txt
altgraph==0.17
fbs==0.8.6
future==0.18.2
macholib==1.14
pefile==2019.4.18
Pillow==7.1.2
PyInstaller==3.4
PyQt5==5.15.0
PyQt5-sip==12.8.0
pywin32-ctypes==0.2.0
